requirements.txt
openai
requests
fitz  # PyMuPDF
chromadb
gradio  # or streamlit depending on the front-end
python-dotenv
